+++ tests/testthat/test-hardhat_parameters.R
@@ -217,21 +217,6 @@ test_that("fit uses config parameters mix from config= and ...", {
 
 })
 
-test_that("fit works with entmax mask-type", {
-
-  rec <- recipe(EnvironmentSatisfaction ~ ., data = attrition[ids, ]) %>%
-    step_normalize(all_numeric(), -all_outcomes())
-
-  expect_no_error(
-    tabnet_fit(rec, attrition, epochs = 1, valid_split = 0.25, verbose = TRUE,
-                      config = tabnet_config( mask_type="entmax"))
-  )
-  expect_no_error(
-    predict(tabnet_fit(rec, attrition, epochs = 1, valid_split = 0.25, verbose = TRUE,
-                       config = tabnet_config( mask_type="entmax")), attrition)
-  )
-
-})
 
 test_that("fit raise an error with non-supported mask-type", {
 
@@ -240,7 +225,7 @@ test_that("fit raise an error with non-supported mask-type", {
   expect_error(
     tabnet_fit(rec, attrition, epochs = 1, valid_split = 0.25, verbose = TRUE,
                       config = tabnet_config( mask_type="max_entropy")),
-    regexp = "either 'sparsemax' or 'entmax' as"
+    regexp = "either 'sparsemax', 'sparsemax15', 'entmax' or 'entmax15' as"
   )
 
 })