+++ NEWS.md
@@ -2,8 +2,10 @@
 
 ## New features
 
+* add optimal threshold and support size into new 1.5 alpha `entmax15()` and `sparsemax15()`
+  `mask_types`. Add an optional `mask_topk` config parameter. (#180)
 * tabnet is now using the `torch_ignite_adam` when available. 
-  Result is 30% faster pretraining and fitting tasks.
+  Result is 30% faster pretraining and fitting tasks (#178).
 * add `nn_aum_loss()` function for area under the $Min(FPR,FNR)$ optimization for cases of 
   unbalanced binary classification (#178).
 * add a vignette on imbalanced binary classification with `nn_aum_loss()` (#178).