+++ NEWS.md
@@ -2,10 +2,10 @@
 
 ## New features
 
-* messaging improved with {cli} 
+* messaging is now improved with {cli} 
 * add optimal threshold and support size into new 1.5 alpha `entmax15()` and `sparsemax15()`
   `mask_types`. Add an optional `mask_topk` config parameter. (#180)
-* tabnet is now using the `torch_ignite_adam` when available. 
+* `optimizer`now default to the `torch_ignite_adam`  when available.
   Result is 30% faster pretraining and fitting tasks (#178).
 * add `nn_aum_loss()` function for area under the $Min(FPR,FNR)$ optimization for cases of 
   unbalanced binary classification (#178).
@@ -13,7 +13,7 @@
 
 ## Bugfixes
 
-* fix config parameter merging issues when parameter is a torch loss or torch optimizer generator.
+* config parameter now merge correctly for torch loss or torch optimizer generator.
 * `nn_unsupervised_loss()` is now a proper loss function.
 
 # tabnet 0.7.0