+++ NEWS.md
@@ -1,4 +1,4 @@
-# tabnet (development version)
+# tabnet 0.8.0
 
 ## New features